nlp/requirements.txt
flask==3.0.2
flask-cors==4.0.0
spacy==3.7.4
python-dateutil==2.8.2
gunicorn==21.2.0 